Basic setup for rendering an image
In the handle_no_event we call the render function which will set the
colors for the pixels and push the image to the window. Right now for
testing it just has a pixel traversing the screen from left to right
wrapping around the edge. The idea is that the render function will loop
over all the pixels in the screen and call another function which will
do all the raytracing. This way it is set up to be almost like a
fragment shader.