per_pixel func for rendering and basic vector func
The render function now loops over all the pixels each time calling the
new function per_pixel. This function is where the raytracing stuff will
happen.

The basic vector functions added are not used yet (and are yet to be
tested) but will be important for later. Since C does not really have
vec3 like the names suggest, we use the t_coordinates structure and
consider them to be the endpoint of the vector with an origin at axis
origin.