Less variables in render, small change pixel_put
Instead of using two variable i and j to know the pixel coordinates and
then setting them in a third variable p, we can directly use the
variable p.

fast_pixel_put changed to set the color in one line instead of three as
it was unnessecary.

In per_pixel I've tested a bit of remapping the uv between -1, 1 to see
if the colors match the way shadertoy and conventional solutions do it.
It seems that we have to explicitly make sure the color channels are
between 0-255 as it would not give the correct black third quadrant
otherwise.