Big speedup adding compiler optimization
After some testing with a profiler, the program seems to be almost
exclusively inside the per_pixel and fast_pixel_put functions. This is
expected and a good sign that it isn't losing time elsewhere. But if
these functions are slow, drawing the image is slow. I've learned that
clang does not optimize code by default and we enable the most reliable
setting for it with the -O3 flag. This has drastically reduced
frametimes.

I've also removed the check for on_screen as we will expect the inputted
pixel to be on screen anyways as this will be part of the raytracing
algorithm, but arguments can be made to make sure we don't cause a
segfault and do still need this check.