Added first basic raytracing example
A simple origin centered sphere is ray-cast inside the per_pixel
function. An aspect ratio is needed to correct the ray direction vector
otherwise the sphere is stretched. This example is just to get a sense
of how the math works in case for a sphere to better understand what
steps are needed and what can be put in seperate functions and what not.
To give the sense of it actually being a 3d world and not a flat circle
on the screen. I've mapped the returned color according to the hit
distance. Closer hits are a bit more white than further ones, and misses
are full black.