Added first example for light shaded sphere
The code iterations for the per_pixel are in if clauses so you can
change to see and test other things. Later these will be gone of course.

I've added two us such blocks. First one calculates the normal vector at
the hitpoint on the sphere and maps it to the color range. This is a
classic example often used to know what the normal vector values are on
an object in 3d space. The second block introduces an infinetly far away
light (like a sun) using a light direction vector. Using the dot product
between the normal and the negative of this light_dir vector we can
shade the sphere. Notice that we don't want negative values.

This along with previous commits give a pretty good understanding of
what the program will have to do. Next steps are to convert this into
neat and clean code and callable functions. After which we can start to
see if we can use the data from the scene read from a file as well as
support to two other objects (plane and cylinder).