Implemented the basic idea looping over objects
per_pixel now calls two new functions. First one casts a ray into the
scene to determine if it hits something. If had did not hit something,
you know that by testing the new struct t_hit_info. If the distance
attribute is inf (using isinf function from math.h) then no object was
hit. If it has hit something, hit_info holds the distance and
coordinates of the hit as well as the object type and its index in the
scene array. This was needed because we need different functions for
each object.

After finding out if a hit occured, we want to calculate the color of
the pixel. The function color_from_hit should handle this by looking at
the hit_info and determine the color according to the rules and physics
math stuff. Many of which is still not implemented. Currently just
calculates the normal and based on the lightpoint's direction shades the
pixel from the sphere's color (albedo). I believe this function will
later have to cast more rays called secondary rays to determine more
info on how to color the pixel.