Added rendering of planes
The orientation vectors are now normalized when being parsed. Without
this, the planes' colors don't show correctly. This is also just
generally what you should do with direction vectors (though we can get
away with not normalizing the ray direction at the moment since it
technically just scales the distance value).

Intersection for the planes is analytically solved. The value of the
denominator in the equation can go to zero when the ray is parallel with
the surface of the plane. Therefor a check is used to see if it is close
to zero or not. We are interested in showing the plane if our ray hits
it on the side the normal vector is pointing out of, so the denominator
should be negative in this case (dot product).