Added cylinder as an object
For this I added a disk struct so this can be used as the caps of the
cylinder. Some floating point rounding is still a bit weird so some
checks have to be done with some small delta in mind.

Solving the quadratic equation is now a seperate function. Added a basic
cornell box setup.