Added a bunch of event handling stuff
These events are inmplemented to allow for movement and rotation of the
camera in the world during program execution. At the moment no rotation
is possible, but the events are set up for them to be implemented.

Added keyboard press and release events for movement of the camera.
Currently moves along world space axes, future iteration will move
according to camera space.

Added mouse button press and release events. Right click will be used as
a click and drag to rotate the camera view. Upon clicking, the mouse
position is stored. During the handle_no_event looping function, if the
right click is held down, new mouse positions is requested. This will be
used as rotation angles, but for the moment translate the camera along x
and y in world space.

For a smooth experience, when a key or button is pressed, a single bit
is set. Each of the possible inputted keys is a single bit in the
variables stored in the mlx_data structure. Thus these attributes hold
the state of each key, either being pressed or not. This state is used
in the handle_no_event loop to decide and calculate what to do.

Upon release the bit is set back to 0. With autokeyrepeat on, a
KeyRelease and KeyPress event is sent out for each repeat. With these
two handler functions set up this way the keys can be held down,
multiple at the same time, resulting in smooth user control.

The 'space' key is used to pause rendering. With the event handling set
up this way, holding the 'space' key results in quickly setting and
unsetting it's state, thus quickly pausing and un-pausing rendering.
This is because of the autokeyrepeat and the fact that we want the key
to behave as a toggle (thus holding should not work as movement keys). A
possible solution is to turn off the autokeyrepeat, but might lead to
other unwanted behaviour so some research is still being done.