Camera can now be rotated
Theres is still some wonky stuff. Moving after having rotated does not
respect the new camera viewpoint. Also, a lot of extra attributes were
added to the camera class as prep for a different approach to rotating
the view. Either we implement that approach or clean up a bunch.

Rotation is achieved by using quaternions which store a rotation around
an axis and can be combined with other quaternions to describe multiple
different axis rotations as one around another axis. Currently mouse
position delta is mapped to pitch and yaw delta used to make a
quaternion to rotate the axes of the camera.