Now rendering at lower res during movement
To make interacting better and smoother, rendering is first done at a
lower resolution by grouping pixels according to dx and dy (currently
both hardcoded as 5). Pressing 'r' renders the scene at full window
resolution and then stops rendering. Moving position or rotation returns
the resulotion to low. This is done by holding a state set by an enum as
LOW, HIGH, DONE.

Pressing 'r' only registers when no movement or mouse button is held
down. This is to prevent certain cases where you hold a key, then
render, let go of the key, and toggle back to low resolution.  The
keystate for the key you held down is now wrongly toggled making you
move while not pressing keys. This issue still occurs when doing the
sequence but moving the window. Possible fix is to not toggle input
states, but explicitly set some values to true or false. Maybe some
function to set a bit to one or zero can be implemented as well.

Also fixed is a rare case where camera position would end up at inf for
all coordinates. A simple check to see if the norm of the move_dir is
zero, prevents dividing by 0.