Added moving and rotating scene objects
Left clicking selects an object (de-selecting the previous one if one
was selected). When an object is selected, camera movement keys are used
to move the object along world axes (might change it to follow camera
axes?). Rotating also uses right click drag, standard yaw and pitch
control, but when left control is held down horizontal mouse movement
controls the roll instead of the yaw.

Many of these things are a bit preferetial settings and can be changed
if we want to. Maybe we can combine all of them with a toggle for which
axes to use?