Caps now actually working
The diameter of the cap is calculated based on the angle. We will have
to change some things. Either take the angle as an argument, or
calculate the angle based on height and base diameter. Right now the
angle is arbitrarily chosen and hard coded.

As for the normal instead of '1e-4' I had typed '1e4' so the check was
never correct to determine if the hit point was on a cap or not.

I think there is still something weird about the normal on the cone
surface. I believe right now it is the same as a cylinder which I don't
think is correct given the angled surface.

Another thing that seems to happen is incorrect hit detection when
almost parallel to the axis of the cone. Best checked by looking at
either the tip or the base head on. Some hit's distance might not be
correct in such cases? Maybe because ray origin is inside the cone?
Other object right now ignore 'hits' if it is not on the side of the
outward normal.