A lot of setup for a menu
Two new structs. One is a viewport ui element, in the mlx_data struct we
hold two of those to be able to switch between fullscreen and only a
part of the screen (when menu is shown). The menu will have some more
stuff to handle buttons? slides? textboxes perhaps? Right now the size
of the menu is set to 300, but should be calculated upon initialization,
resizing is not in the works yet, but it would require releasing an
reallocating a new image every time.

The menu opens when an object is selected. Open or close is with 'm'
key. This key is ignored if we are rendering or have rendered in full
resolution. Menu is closed automatically when full res was activated.

When the menu is open, a simple check is ran to see if a click happens
within its bounds or not.

This is a rather simple test to see if it would be possible.