No redraw after menu has been drawn
Similar to the full_res check and values, the menu can be either set to
not show, wait to be drawn on screen or has been drawn. This removes
updating and redrawing pixels that do not change.

Also includes a function to initialize an image as it repeats the same
code.