Added a lot regarding setting up menu ui elements
At this moment the sturcture is that a menu points to an heap allocated
array of pages. Each page has its own attributes, most notable is a heap
allocated array of elements. Each element is made to point to a draw
function and a function should if have some function.

This means that we (handcrafted) make multiple pages, each showing
different elements and the elements themselfs know how to be drawn so we
can just loop over them.

This push is to have something functional as a proof that it could
potentially work. There is still a lot to be done. First change I want
to make is the positions of the page and elements to not be relative to
their parent as it requires knowledge of the parents position which
shouldn't be necessary.