Page and elements dont need parent pointer anymore
Made the element draw function take a position which also removes the
need for it in the struct. It might come back later for when we check if
clicks happen within some bounds.

This change was made possible by not holding a position relative to
parent, but to the screen.