First working button to reset camera
This shows that we can click on elements in the list on the page,
activating the function if there is one.

I am not sure if I like this code flow, it is not as self-working as I
had hoped for. Every element will have to point to its own function as
it can't be generalized I think. Each function will take t_mlx_data* to
have acces to pretty much everything and only the function body will
know what to do. Would have loved a more reusable code structure,
generic functions that know what to do by looking at the element self
pointer.