key_release calls handle_movement_keys, cond order
Realized the code which handles key release events for the movement keys
is exactly the same as the press events (because it is just a toggle of
a bit). Thus it can just call that function.

I reordered the 'r' key condition checks to first check if it is a 'r'
or not. I am uncertain which order is faster/better.