More correct light and gamma correction added
Since light from a point source travels outward from said point, it
loses intensity proportional to the 1 over the distance squared. This
follows the inverse square law. It is then also a lot darker given the
size of the scene and objects. Thus we will have to scale them down in
the .rt file. This is done for the cornell_box .rt file but not for the
others as we might implement it differently.

The color of the light source is now also accounted for. Setting the rgb
values of the light to not be close to white, we see different colors on
objects and some being black. This is because the color setting is
actually the albedo, which determines how much of each component is
reflected as a ratio (between 0 and 1). Thus if a light source has no
component for the red color, and an object has only a component in the
red channel. Then the light reflected from that object will be
black. The light is fully absorbed.

Also included is color correction from linear rgb space to srgb. One can
either do it fast by taking sqrt of the channels, or follow a more
expensive formula.