Fixed cone noisy base and wrong hit when inside
The noisy base was due to floating point precision, a small delta is
used.

When the ray origin was inside the infinite cone, the wrong hit was
being used as solve_quadratic was written to return the closest
non-negative hit. But because we want to discard the negative side of
the cone, we want the furthest hit in this situation.

I've fixed it by first checking if the ray origin is inside the infinite
cone and if it is the case, checking if it is in the positive or the
negative side. The positive side just wants to hit the base cap, the
negative want the furthest hit. As a simple solution, this last case
just calls 'solve_quadratic2' which returns the furthest non-negative
hit.

This is probably not the best way and it makes the function too big for
norminette. I'll be looking for a better implementation of the fix.
Maybe solve_quadratic can return both hits and the intersection
functions will figure out what to do.