Added shadow detection
The function 'cast_shadow_ray' is a simplere version of the
'cast_ray' function as it may return with a true or false. It only has
to detect if a point in space is obstructed by an object, or if light
from a point light can reach it. A ray is cast from the point to the
light and checks for the obstructions. Hits that lie beyond the light
must be ignored, thus we pass the distance from our point to the light
as an argument.

We have to prevent objects casting shadows on themselfs because of
floating point precision error. This is called commonly called shadow
acne as it produces noise like patterns. Often, the solution is to move
the shadow ray origin ever so slightly along the surface normal of the
object. I've also seen succes with only casting shadows if the dot
product of the ray direction and the normal is positive. If it were
negative, the ray would have to travel into the object. This is also
less expensive than to loop through the scene again. Both solutions have
been implemented but have to be switched between using the #if block, to
choose later on what is best.