Full phong shading added
Phong shading is a sum of three terms: ambient, diffuse and specular.
The diffuse is the same as we had previously, using lambert shading (and
taking in to account the inverse square law though I'm not sure if we'll
keep that in). The ambient term is easily calculated as the object
albedo times the color intensities of the ambient light (as entered in
the '.rt' file). The specular reflection is not multiplied by the
objects albedo. This should actually be changed depending on the
material. A dielectric will reflect the color of the light source, a
conductor (metal) material will absorb a portion thus has to multiplied
by its albedo.

The constants k_a, k_s and k_d indicate how much of the terms add to the
color of the pixel and should ideally be on a per-object basis stored in
a meterial structure. For testing convenience, in this push, they are
just set inside the phong_shading function.