Basic setup for a menu using displays.
In an earlier branch, I've tried to make a setup for a menu all inside
the same window. This proved to be do-able, but very annoying to
implement. Xlib windows can help with that, since each window can have
its own hooks, events and hanlder functions.

Now to make this possible within this project, we will have to deviate
from only using minilibx. This is because the implementation that is
written, sets a lot of values for us. It makes it easier to get a basic
program, but for what we want to do, we want more control over what
values are used and set. We will basically have to understand and use
Xlib ourselfs without using mlx.

This push serves as an example for that, by using a couple of Xlib
functions we can create more what we desire than with pure mlx. First,
when mlx creates windows, it always chooses the XDefaultRootWindow as
parent window. This has as effect that every window will actually be a
seperate window, each drag-able over the screen. What we want is for
mulitple windows to be contained inside one, one drag-able window. For
this we set the parent window ourselfs.

Then for the ability to have a viewport, and after pressing switch to a
renderport. We have to map and unmap the window we want to show or
un-show respectively. When a window is unmapped it will also not send
events, which is usefull in its own way, but is the reason we have to
hook some events for the renderport as well.