More advanced setup for menu, with side and bottom
Sizes of the different menu parts are calculated based on the size of
the screen, menu bar height and the side menu width. This feels like a
more advanced interface many programs adopt. Next plan is to make the
different buttons and have multiple side menus, each for a different
type of object of interactable. Camera, light, etc.. Still requires a
lot of work.