Merge phong_shading into main
A couple of conflicts had to be resolved. For now I also have added the
small bias along the surface normal to avoid selfcolision causing shadow
acne. I had prevously fixed it for spheres by only allow hits in one
way. But this method would not work for cones. Now when you are inside a
sphere, you will not see through it and instead the the shade of the
sphere.