Some norminette for shading and casting
For the shading I could split it in two functions, though I don't quite
like how it looks, it does work for the norminette.

As for the casting of the ray as well as the shadow ray checks, I had to
do them in two different files. Reason is that I had to make a function
for each object to loop over the objects of that type. If we add another
object type, we will again have to split the files as they now have five
functions which is the max allowed according to the norminette. A better
approach could have been achieved with a bit more object oriented
thinking. Creating a structure that itself knows what object type it is
and which functions it should call for certain things by using function
pointers.