Rotating and moving now norminette
As part of user interaction with the program, one can rotate and move
objects and the camera. This is now norminette compliant. Moved the
functions in a file, did some finessing. Now each object utils file has
a function for move and rotate which is called when needed. They do the
same thing for each though.