Norminette for mlx_events, move and rotate
Moving and rotating camera and objects now is seperate file called in
the mlx_no_event.c file. In this file the main loop of the program is
run. Handling keyboard inputs is done in seperate file, as well as the
other events for mouse button presses and the destroy event.

Added a couple of handy features where you can hold shift to rotate/move
faster or control to go slower. A global speed can also be increased or
decreased with the up and down arrow keys respectively. This gives more
control to the user and less dependence on scale of the scene for how
fast you feel.