Better low res support, ignore keys when full_res
Previously edges where not always correct. Either not enough pixels
where drawn because the sample pixel was beyond the screen size. Or the
sample pixel was on the right edge of the screen, bleeding its colors to
the left edge (as it is one continuous memory). Fixed by letting the
sample pixel loop go one dx and dy past the size of the screen and
adding a check in the set_pixels loops for the filled rectangles so it
only tries to put pixels that are within screen bounds.

Also includes a simple check to not allow for some key inputs if the
program is not in the interactive low_res mode. Just by returning in the
key_down event handler.