Added more function to number box
Elements now have a void* data which is set to point to an attribute of
an object, camera or similar. The box now dynamically creates a string
to display the value of said data. To know what kind of data it holds,
check the data_type attribute.

Currectly have implemented the ability to click a number box in hopes
one can type and manipulate the attributes more directly. At the moment
it just colors the text magenta to know that you can at least select
them.

This way of gathering input can be beneficial as you don't require to go
to the terminal to type something in. Reading from the terminal also
requires a lot of care in how to handle key inputs still to the
rendering window. In this branch, we'll be just checking if a textbox is
selected and handle keypresses differently in that case. Bonus to handle
auto key repeat correctly.