Added taking keyspresses for numberboxes
Now after selecting a box, the keystrokes are handled in a different
function. It allows for number input '0'-'9', space ' ', comma ',',
period '.', minus '-', plus '+'. You can also use the backspace key to
remove the last character in the string. With the enter key, you apply
the change to the attribute linked to the box and exit this editing
state. One can use the escape key to exit it without applying the
change.

Clicking another menu element while in this editing state you exit the
current selected one without applying the inputted string.