Added yaw and pitch boxes and apply functions
Now each element has an apply function pointer. This is to have
different behaviour depending on what box you put info in. For a simple
coordinate, not much has to be done. But for pitch and yaw, some checks
on the value need to be done as well as updating the rotation quaternion
and camera vectors.

User can now also input text. Prepending the yaw and pitch input values
with a 'd' interpretes them as degrees instead of the default radians.
Could also code it to be the other way (default degr, prepend 'r' for
radians). Check on pitch value is done to be in range of accepted values
[-pi/2, pi/2].