pitch more logical, nbox color matches input
Small change for the way the pitch is stored and used. Now feels more
natural to think about the rotation, positive is looking up, negative is
looking down.

Number box color for it's text now also changes if it is currently
catching input keys. Changed this color to match the clickable button
elements. Colors could be changes to preference later.