Removed need of separat apply func attribute
Recent push added this attribute to the ui_element structure. It is not
really needed as the on click function for a number box does not have to
be a separate/different function. All the on click for a box has to do
is to set up the current input attributes.