menu elements for light added
Ambient light has boxes for intensity and color. Point light has them
alongside the position coordinates. As this branch has an incomplete
renderer, the intensity and color of the lights are not used, there is
not even support for ambient light at the moment. So some things can't
be fully tested yet.

Input for color values can be as a float in range [0, 1]. To input an
integer value [0, 255], prepend with 'i', to input using hex value,
prepend with '0x'.

Input for intensities can be as a float in range [0, 1] representing a
percent. Prepend with a 'p' to input percentage between [0, 100].