Basic pages for objects and fixed leaks
When selecting an object, the menu elements for the object should be
linked to that object. To not have to rewrite the 'same' code in two
different places (would be a pain to match up every time we change the
layout and is prone to bugs), I wrote it in such a way that elements are
freed, allocated and set each time an object is selected.

I had some problems with this approach, causing invalid frees for the
element lists. This was resolved by making the curr_page attribute of
the menu be a pointer. I've also taken the time to fix several memory
leaks introduced by the development of the menu. In ft_ftoa a leak
happend by not freeing the previous string allocated to the result. At
the end of program execution, now the menu pages and their elements are
freed as well as the singleton mlx_data allocated in main.