Added more menu elements for objects
Now each object has coordinates and rgb along with their attributes like
diameter and height.

Added fov number box in home page for the camera. At the moment this is
the only one using the UI_DATA_INT, all the other boxes are floats.
Seems to work.

Also now starting the render with 'r' does not try to hide the menu.
This makes it so returning to interactive mode (again 'r') respects the
state of the menu it was in before starting the render.