Page elements are now as a list
They were previously a dynamically allocated array, which has its pros.
But it made some other things a lot more annoying. With the approach to
make them a linked list, we solve a couple of problems:

Making pages is a lot more user coder friendly, no need to count the
amount of elements and a lot less manually setting values.

Creating pages is now more dynamic. If in the future we are able to
add or remove objects from the scene, a object select page is now easier
to make.

It is easier to create group elements. For example the three coordinates
of an object. They can be grouped in one function that creates all
three.

Making the code norminette friendly is now a lot more doable.

Clicking and getting the right element is still as easy, the index is
calculated, and the list traversed that many step. Simple, right?