Added the ability to slide the number boxes
As another method of changing the values shown by number boxes, you can
now click-and-draw on them left an right to change its associated value.
For this, I added a couple of attributes to the element structure
(making it harder to work with norminette wise). A min and max value for
the data linked to the box as well as a step value indicating how much
change is done for every pixel moved with the mouse.

Upon selecting a number box, the text is not immediately replaced by the
inputted string. Instead only after having typed something (a.k.a. the
string is not NULL), does it's text change on screen.

Also added a simple ft_calloc for when backspace is the first key after
selecting a number box. This makes it so shown text is empty,
effectively starting with the empty string. Pressing enter results in
trying to pass a value of zero. The ft_calloc does not specifically have
to be checked for a malloc failure as nothing seriously bad happens when
the string is NULL.