Added object select page and home button
Every page (except home itself) now has a button to go to the home page.
Home page now has a button to go to the object select page. This page
lists every object currently in the scene as a button. Clicking on an
object button selects that object as if it were clicked via the rendered
viewport.

Something I don't quite like is that the string for each object button
could not be allocated without creating a leak. This is because other
elements' strings are not necesarilly allocated, and no way of knowing
which needs to be freed is implemented yet. Could be as easy as creating
a separate struct holding a char* str and a bool saying if it should be
freed or not. Other option is to always allocate regardless, but that
brings some stuff with it I don't really want to deal with.