Added and object add and object delete page
The add page lists buttons, one for each object supported by the
program. Upon clicking, it adds a default version of that to the scene.
Because the scene is an allocated array, a new array has to be
allocated, filled with the current data and the old one freed. It
automatically selects the object created, with the menu going to that
page. The select and delete pages are updated with an extra button for
this new object.

The delete page does something similar. It lists the objects in the
scene, (like the select page) and removes them when clicked. Again a new
array is allocated, filled with the correct scene data by skipping the
selected object. The menu stays open in case the user wants to remove
more objects.