Menu elements now scrollable
When user scrolls in the menu region, the elements are shifted up or
down. They are only shifted in a direction if there are hidden elements
in that direction. To show that there are elements hidden, a hint is
shown via a '^' at the top and a 'V' at the bottom.

I've made sure that clicking the buttons still work after being shifted.

This is gonna be so much fun to norminette ... not!