Big progress on norminetting the menu stuff
Progress tracked in menu_norm_stuff.txt. Pretty much only the pages have
toe be done. They have just so many lines, at least one per element on a
page.

To adhere to the rules of the norm, I had to introduce a couple of
really stupid solutions. Example is a couple of extra structures just so
that the amount of arguments did not exceed four.

Also fixed an issue when clicking on the number box after having typed
in int resulting in an invalid free.