Good progress on norminette after menu merge
Some functions had a lot more lines than allowed, many of which because
of the integration of the menu. The pages have a lot of elements, each
needing at least one line to be created. For the repeating parts, I've
grouped them together into three separate functions: rgb, pos,
direction. The direction function takes an extra boolean argument to
decide which reset function to use, objects want to reset along the y
axis, while the camera want to align with the negative z.

I've also fixed the norm found in the parse_camera.c file.

Applied the logic for the rotation of the camera introduced in the
docked_menu branch. This just means the displayed angle makes a bit more
sense.

Next thing to work on is the event files, the menu has introduced some
extra lines, making them to long. Many of the keybuttons could be
replaced by a numberbox in the menu, though having both might be a plus.