Added cone to menu and some norm stuff
Cone is fully integrated into the menu. Can add, select, delete as well
as interact with the attributes of the cone via its page.

Had to put the initialization of the select and delete page in a
separate file for norm reasons, too many functions in a single file
otherwise.

Also changed the direction of scrolling over the menu, I knew something
was off.