More norm stuff, bit of error checks, dark page bg
In general more norminette stuff. A couple of error handling changes.
Initialization of the pages happens with malloc which sets errno,
checking for that variable lets us know if something went wrong.

Added an extra img in the menu struct to have a bit of a contrasting
background for the page itself. Also implemented freeing it.

I had hoped to get rid of the usleep in init_menu by not using memset to
set the color of the background images. But it did not seem to work. I
still do not know what the cause is for the color not being correct at
times, and only the very first time we draw them to the screen. It is
almost as if the image is just slow to load or something, that's why a
usleep would work.