Norminette the cone utils
Most work was the cone_hit function, but wasn't all that bad after
recognizing the similar works for the quadratic solve and the cap hits.
The difference was in what conditions which function had to be called.
The in_inf boolean holds true when the ray origin is inside the infinite
sized, quadratic cone (without one side culled). Depending on that value
the correct solve_quadratic function is called.