Fixed cam pitch not behaving correctly
When changing the pitch of the camera, there was a check to see if the
resulting pitch was outside the allowed range [-pi/2, pi/2]. This was
calculated with positive diff but the actual change was done with
negative diff. The negative is now done before checking.

When the user did fast movement, sometimes a diff bigger than 1 pixel is
captured. With the previous check, diff was set to zero resulting in an
incorrect early stop of the pitch. Now it sets the value to the min or
max, essentially a clamp.